+++ lib/auth/view/find_pw_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:psr/auth/component/input_user_info.dart';
 
 import '../../common/layout/default_appbar_layout.dart';
 import '../../common/layout/purple_filled_button.dart';
 import '../component/complete_reset_pw.dart';
-import 'signup/input_user_info_screen.dart';
 import '../component/reset_password.dart';
 import 'login_screen.dart';
 
@@ -18,7 +18,7 @@ class _FindPWScreenState extends State<FindPWScreen> {
   bool isInfoInputted = false;
   bool isReset = false;
 
-  final TextEditingController nameController = TextEditingController();
+  final TextEditingController idController = TextEditingController();
   final TextEditingController phoneNumController = TextEditingController();
   final TextEditingController codeController = TextEditingController();
 
@@ -26,7 +26,7 @@ class _FindPWScreenState extends State<FindPWScreen> {
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.white,
-      appBar: const DefaultAppBarLayout(titleText: '비밀번호 재설정',),
+      appBar: DefaultAppBarLayout(titleText: '비밀번호 재설정', isBackItemHidden: isReset,),
       body: renderBody(),
       bottomNavigationBar: PurpleFilledButton(
         title: (isReset) ? '로그인' : '다음',
@@ -42,10 +42,15 @@ class _FindPWScreenState extends State<FindPWScreen> {
     } else if (isInfoInputted && !isReset) {
       return const ResetPassword();
     } else {
-      return const Column(
+      return Column(
         children: [
-          SizedBox(height: 40,),
-          InputUserInfoScreen(),
+          const SizedBox(height: 40,),
+          InputUserInfo(
+            isTopName: false,
+            nameController: idController,
+            phoneNumController: phoneNumController,
+            validCodeController: codeController,
+          )
         ],
       );
     }
